Fix progress bar padding issue

+++ Source/ScaredyImp/Private/UI/HealthBarWidget.cpp
@@ -6,7 +6,7 @@ void UHealthBarWidget::SetHealth(int32 CurrentHealth, int32 MaxHealth)
 {
 	if (!IsValid(HealthProgressBar)) return;
 
-	const float HealthPercent = MaxHealth > 0 ? CurrentHealth / MaxHealth : 0.0f;
+	const float HealthPercent = MaxHealth > 0 ? static_cast<float>(CurrentHealth) / static_cast<float>(MaxHealth) : 0.0f;
 
 	HealthProgressBar->SetPercent(FMath::Clamp(HealthPercent, 0.0f, 1.0f));
 }
